# openapi/membership.yaml
openapi: 3.1.0
info:
  title: Sports Club Management - Membership API
  version: 1.0.0
  description: |
    Membership application and approval endpoints for the Sports Club Management System.
    Handles application submission, review, approval/rejection workflows.
  contact:
    name: API Support
    email: support@sportsclub.example.com
servers:
  - url: https://your-app.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server (not used in production-only workflow)
...
paths:
  /membership/apply:
    post:
      tags:
        - Applications
      summary: Submit membership application
      description: |
        Submit a new membership application for a specific club. Creates application
        record with personal information and documents. Updates profile membership_status
        to 'pending'.
        
        **Idempotency**: This endpoint supports idempotency via the `Idempotency-Key` header.
        Duplicate requests with the same key will return the cached response.
        
        **Validates: Requirements 2.1, 2.2, 20.6, 20.7**
      operationId: submitApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CausationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSubmissionRequest'
            examples:
              completeApplication:
                summary: Complete application with all fields
                value:
                  club_id: 550e8400-e29b-41d4-a716-446655440000
                  personal_info:
                    full_name: สมชาย ใจดี
                    nickname: ชาย
                    date_of_birth: "2010-05-15"
                    phone_number: "0812345678"
                    gender: male
                    medical_conditions: ไม่มีโรคประจำตัว
                    emergency_contact_name: นางสาวสมหญิง ใจดี
                    emergency_contact_phone: "0898765432"
                    emergency_contact_relationship: มารดา
                  documents:
                    - type: id_card
                      url: https://storage.supabase.co/...
                    - type: medical_certificate
                      url: https://storage.supabase.co/...
      responses:
        '200':
          description: Application submitted successfully
          headers:
            X-Request-Id:
              description: Unique request identifier
              schema:
                type: string
                format: uuid
            X-Idempotency-Cached:
              description: Indicates if response was cached (idempotent request)
              schema:
                type: string
                enum: ['true', 'false']
            X-Original-Timestamp:
              description: Timestamp of original request (if cached)
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSubmissionResponse'
              examples:
                success:
                  summary: Successful submission
                  value:
                    success: true
                    data:
                      applicationId: 660e8400-e29b-41d4-a716-446655440001
                    metadata:
                      timestamp: "2025-11-27T10:30:00Z"
                      requestId: 770e8400-e29b-41d4-a716-446655440002
        '400':
          description: Validation error or duplicate application
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateApplication:
                  summary: User already has pending application
                  value:
                    success: false
                    error:
                      code: DUPLICATE_APPLICATION
                      message: คุณมีใบสมัครที่รอการอนุมัติอยู่แล้วสำหรับ ชมรมฟุตบอล กรุณารอการพิจารณาก่อนสมัครใหม่
                validationError:
                  summary: Invalid input data
                  value:
                    success: false
                    error:
                      code: VALIDATION_ERROR
                      message: ข้อมูลไม่ถูกต้อง
                invalidIdempotencyKey:
                  summary: Invalid idempotency key format
                  value:
                    success: false
                    error:
                      code: INVALID_IDEMPOTENCY_KEY
                      message: Idempotency-Key must be a valid UUID or alphanumeric string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Unique key to ensure idempotent request processing. Use UUID or alphanumeric string.
        Duplicate requests with the same key return cached response.
      required: false
      schema:
        type: string
        format: uuid
        example: 550e8400-e29b-41d4-a716-446655440000
    CorrelationId:
      name: X-Correlation-ID
      in: header
      description: Correlation ID for request tracing across operations
      required: false
      schema:
        type: string
        format: uuid
    CausationId:
      name: X-Causation-ID
      in: header
      description: Causation ID linking cause and effect operations
      required: false
      schema:
        type: string
        format: uuid
  schemas:
    PersonalInfo:
      type: object
      required:
        - full_name
        - date_of_birth
        - phone_number
        - gender
      properties:
        full_name:
          type: string
          minLength: 2
          maxLength: 255
          description: Full name in Thai or English
          example: สมชาย ใจดี
        nickname:
          type: string
          maxLength: 50
          description: Preferred nickname
          example: ชาย
        date_of_birth:
          type: string
          format: date
          description: Date of birth (must be at least 5 years old)
          example: "2010-05-15"
        phone_number:
          type: string
          pattern: '^0\d{9}$'
          description: Thai phone number (10 digits starting with 0)
          example: "0812345678"
        gender:
          type: string
          enum: [male, female, other]
          description: Gender
          example: male
        medical_conditions:
          type: string
          maxLength: 1000
          description: Medical conditions or health notes
          example: ไม่มีโรคประจำตัว
        emergency_contact_name:
          type: string
          maxLength: 255
          description: Emergency contact person name
          example: นางสาวสมหญิง ใจดี
        emergency_contact_phone:
          type: string
          pattern: '^0\d{9}$'
          description: Emergency contact phone number
          example: "0898765432"
        emergency_contact_relationship:
          type: string
          maxLength: 100
          description: Relationship to emergency contact
          example: มารดา
    Document:
      type: object
      required:
        - type
        - url
      properties:
        type:
          type: string
          enum: [id_card, medical_certificate, other]
          description: Document type
        url:
          type: string
          format: uri
          description: Signed URL to document in storage
          example: https://storage.supabase.co/object/sign/membership-documents/...
    ApplicationSubmissionRequest:
      type: object
      required:
        - club_id
        - personal_info
        - documents
      properties:
        club_id:
          type: string
          format: uuid
          description: UUID of the club to apply to
          example: 550e8400-e29b-41d4-a716-446655440000
        personal_info:
          $ref: '#/components/schemas/PersonalInfo'
        documents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Document'
          description: Array of uploaded document references
    ApplicationSubmissionResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            applicationId:
              type: string
              format: uuid
              description: UUID of created application
        metadata:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
            cached:
              type: boolean
              description: True if response was cached (idempotent request)
            originalTimestamp:
              type: string
              format: date-time
              description: Timestamp of original request (if cached)
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - AUTHENTICATION_REQUIRED
                - AUTHORIZATION_ERROR
                - NOT_FOUND
                - DUPLICATE_APPLICATION
                - ALREADY_PROCESSED
                - INVALID_FILE_TYPE
                - FILE_TOO_LARGE
                - INVALID_IDEMPOTENCY_KEY
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error message (Thai language)
            details:
              type: object
              description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: AUTHENTICATION_REQUIRED
              message: กรุณาเข้าสู่ระบบ
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: ลองบ่อยเกินไป กรุณารอสักครู่แล้วลองใหม่อีกครั้ง
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60
security:
  - bearerAuth: []